Previous branh deleted due to github client bug.
This commit contans the HuffyInt type and its overloads, the
HuffyManager which creates Huffman tree's based on priority queue
created from frequency maps which the HuffyInt type interacts with when
it has been modified. Also present are a couple of unimplimented types
and an input and output manager which do self explanitory stuff. Theres
also some rough and ready unit tests in Huffy.cpp.